requirements.txt
streamlit
pandas
numpy
fpdf
xlsxwriter
python-dateutil